crew-ai-travelagent/requirements.txt
crewai
crewai-tools
openai
python-dotenv
requests
pydantic
typing-extensions
pydantic-settings
streamlit>=1.37.0